+++ oap-cache/oap/src/main/java/org/apache/spark/sql/execution/datasources/parquet/SkippableVectorizedColumnReader.java
@@ -18,21 +18,24 @@
 package org.apache.spark.sql.execution.datasources.parquet;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.TimeZone;
 
 import org.apache.parquet.bytes.ByteBufferInputStream;
 import org.apache.parquet.bytes.BytesInput;
 import org.apache.parquet.bytes.BytesUtils;
 import org.apache.parquet.column.ColumnDescriptor;
+import org.apache.parquet.column.Dictionary;
 import org.apache.parquet.column.Encoding;
-import org.apache.parquet.column.page.DataPageV1;
-import org.apache.parquet.column.page.DataPageV2;
-import org.apache.parquet.column.page.PageReader;
+import org.apache.parquet.column.page.*;
 import org.apache.parquet.column.values.ValuesReader;
+import org.apache.parquet.io.api.Binary;
 import org.apache.parquet.schema.OriginalType;
-
+import org.apache.parquet.schema.PrimitiveType;
 import static org.apache.parquet.column.ValuesType.REPETITION_LEVEL;
 
+import org.apache.spark.sql.catalyst.util.DateTimeUtils;
+import org.apache.spark.sql.execution.datasources.SchemaColumnConvertNotSupportedException;
 import org.apache.spark.sql.execution.vectorized.WritableColumnVector;
 import org.apache.spark.sql.types.DataType;
 import org.apache.spark.sql.types.DataTypes;
@@ -42,7 +45,37 @@ import org.apache.spark.sql.types.DecimalType;
  * Add skip values ability to VectorizedColumnReader, skip method refer to
  * read method of VectorizedColumnReader.
  */
-public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
+public class SkippableVectorizedColumnReader {
+
+  /**
+   * Total number of values read.
+   */
+  protected long valuesRead;
+
+  /**
+   * value that indicates the end of the current page. That is,
+   * if valuesRead == endOfPageValueCount, we are at the end of the page.
+   */
+  protected long endOfPageValueCount;
+
+  /**
+   * The dictionary, if this column has dictionary encoding.
+   */
+  protected final Dictionary dictionary;
+
+  /**
+   * If true, the current page is dictionary encoded.
+   */
+  protected boolean isCurrentPageDictionaryEncoded;
+
+  /**
+   * Maximum definition level for this column.
+   */
+  protected final int maxDefLevel;
+
+  private SpecificParquetRecordReaderBase.IntIterator definitionLevelColumn;
+
+  protected ValuesReader dataColumn;
 
   /**
    * Reference dataColumn, but Type is SkippableVectorizedValuesReader, reduce manual cast
@@ -56,13 +89,483 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
    */
   private SkippableVectorizedRleValuesReader defColumnRef;
 
-  public SkippableVectorizedColumnReader(
-      ColumnDescriptor descriptor,
-      OriginalType originalType,
-      PageReader pageReader,
-      TimeZone convertTz)
-      throws IOException {
-    super(descriptor, originalType, pageReader, convertTz);
+  /**
+   * Total number of values in this column (in this row group).
+   */
+  private final long totalValueCount;
+
+  /**
+   * Total values in the current page.
+   */
+  protected int pageValueCount;
+
+  private final PageReader pageReader;
+  protected final ColumnDescriptor descriptor;
+  protected final OriginalType originalType;
+  // The timezone conversion to apply to int96 timestamps. Null if no conversion.
+  private final TimeZone convertTz;
+  private static final TimeZone UTC = DateTimeUtils.TimeZoneUTC();
+
+   public SkippableVectorizedColumnReader(
+           ColumnDescriptor descriptor,
+           OriginalType originalType,
+           PageReader pageReader,
+           TimeZone convertTz) throws IOException {
+    this.descriptor = descriptor;
+    this.pageReader = pageReader;
+    this.convertTz = convertTz;
+    this.originalType = originalType;
+    this.maxDefLevel = descriptor.getMaxDefinitionLevel();
+
+    DictionaryPage dictionaryPage = pageReader.readDictionaryPage();
+    if (dictionaryPage != null) {
+      try {
+        this.dictionary = dictionaryPage.getEncoding().initDictionary(descriptor, dictionaryPage);
+        this.isCurrentPageDictionaryEncoded = true;
+      } catch (IOException e) {
+        throw new IOException("could not decode the dictionary for " + descriptor, e);
+      }
+    } else {
+      this.dictionary = null;
+      this.isCurrentPageDictionaryEncoded = false;
+    }
+    this.totalValueCount = pageReader.getTotalValueCount();
+    if (totalValueCount == 0) {
+      throw new IOException("totalValueCount == 0");
+    }
+  }
+
+  /**
+   * Advances to the next value. Returns true if the value is non-null.
+   */
+  private boolean next() throws IOException {
+    if (valuesRead >= endOfPageValueCount) {
+      if (valuesRead >= totalValueCount) {
+        // How do we get here? Throw end of stream exception?
+        return false;
+      }
+      readPage();
+    }
+    ++valuesRead;
+    // TODO: Don't read for flat schemas
+    //repetitionLevel = repetitionLevelColumn.nextInt();
+    return definitionLevelColumn.nextInt() == maxDefLevel;
+  }
+
+  /**
+   * Reads `total` values from this columnReader into column.
+   */
+  public void readBatch(int total, WritableColumnVector column) throws IOException {
+    int rowId = 0;
+    WritableColumnVector dictionaryIds = null;
+    if (dictionary != null) {
+      // SPARK-16334: We only maintain a single dictionary per row batch, so that it can be used to
+      // decode all previous dictionary encoded pages if we ever encounter a non-dictionary encoded
+      // page.
+      dictionaryIds = column.reserveDictionaryIds(total);
+    }
+    while (total > 0) {
+      // Compute the number of values we want to read in this page.
+      int leftInPage = (int) (endOfPageValueCount - valuesRead);
+      if (leftInPage == 0) {
+        readPage();
+        leftInPage = (int) (endOfPageValueCount - valuesRead);
+      }
+      int num = Math.min(total, leftInPage);
+      PrimitiveType.PrimitiveTypeName typeName =
+              descriptor.getPrimitiveType().getPrimitiveTypeName();
+      if (isCurrentPageDictionaryEncoded) {
+        // Read and decode dictionary ids.
+        defColumnRef.readIntegers(num, dictionaryIds,
+                column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+
+        // TIMESTAMP_MILLIS encoded as INT64 can't be lazily decoded as we need to post process
+        // the values to add microseconds precision.
+        if (column.hasDictionary() || (rowId == 0 &&
+                (typeName == PrimitiveType.PrimitiveTypeName.INT32 ||
+                        (typeName == PrimitiveType.PrimitiveTypeName.INT64 &&
+                                originalType != OriginalType.TIMESTAMP_MILLIS) ||
+                        typeName == PrimitiveType.PrimitiveTypeName.FLOAT ||
+                        typeName == PrimitiveType.PrimitiveTypeName.DOUBLE ||
+                        typeName == PrimitiveType.PrimitiveTypeName.BINARY))) {
+          // Column vector supports lazy decoding of dictionary values so just set the dictionary.
+          // We can't do this if rowId != 0 AND the column doesn't have a dictionary (i.e. some
+          // non-dictionary encoded values have already been added).
+          column.setDictionary(new ParquetDictionaryWrapper(dictionary));
+        } else {
+          decodeDictionaryIds(rowId, num, column, dictionaryIds);
+        }
+      } else {
+        if (column.hasDictionary() && rowId != 0) {
+          // This batch already has dictionary encoded values but this new page is not. The batch
+          // does not support a mix of dictionary and not so we will decode the dictionary.
+          decodeDictionaryIds(0, rowId, column, column.getDictionaryIds());
+        }
+        column.setDictionary(null);
+        switch (typeName) {
+          case BOOLEAN:
+            readBooleanBatch(rowId, num, column);
+            break;
+          case INT32:
+            readIntBatch(rowId, num, column);
+            break;
+          case INT64:
+            readLongBatch(rowId, num, column);
+            break;
+          case INT96:
+            readBinaryBatch(rowId, num, column);
+            break;
+          case FLOAT:
+            readFloatBatch(rowId, num, column);
+            break;
+          case DOUBLE:
+            readDoubleBatch(rowId, num, column);
+            break;
+          case BINARY:
+            readBinaryBatch(rowId, num, column);
+            break;
+          case FIXED_LEN_BYTE_ARRAY:
+            readFixedLenByteArrayBatch(
+                    rowId, num, column, descriptor.getPrimitiveType().getTypeLength());
+            break;
+          default:
+            throw new IOException("Unsupported type: " + typeName);
+        }
+      }
+
+      valuesRead += num;
+      rowId += num;
+      total -= num;
+    }
+  }
+
+  private boolean shouldConvertTimestamps() {
+    return convertTz != null && !convertTz.equals(UTC);
+  }
+
+  /**
+   * Helper function to construct exception for parquet schema mismatch.
+   */
+  private SchemaColumnConvertNotSupportedException constructConvertNotSupportedException(
+          ColumnDescriptor descriptor,
+          WritableColumnVector column) {
+    return new SchemaColumnConvertNotSupportedException(
+            Arrays.toString(descriptor.getPath()),
+            descriptor.getPrimitiveType().getPrimitiveTypeName().toString(),
+            column.dataType().catalogString());
+  }
+
+  /**
+   * Reads `num` values into column, decoding the values from `dictionaryIds` and `dictionary`.
+   */
+  private void decodeDictionaryIds(
+          int rowId,
+          int num,
+          WritableColumnVector column,
+          WritableColumnVector dictionaryIds) {
+    switch (descriptor.getPrimitiveType().getPrimitiveTypeName()) {
+      case INT32:
+        if (column.dataType() == DataTypes.IntegerType ||
+                DecimalType.is32BitDecimalType(column.dataType())) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              column.putInt(i, dictionary.decodeToInt(dictionaryIds.getDictId(i)));
+            }
+          }
+        } else if (column.dataType() == DataTypes.ByteType) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              column.putByte(i, (byte) dictionary.decodeToInt(dictionaryIds.getDictId(i)));
+            }
+          }
+        } else if (column.dataType() == DataTypes.ShortType) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              column.putShort(i, (short) dictionary.decodeToInt(dictionaryIds.getDictId(i)));
+            }
+          }
+        } else {
+          throw constructConvertNotSupportedException(descriptor, column);
+        }
+        break;
+
+      case INT64:
+        if (column.dataType() == DataTypes.LongType ||
+                DecimalType.is64BitDecimalType(column.dataType()) ||
+                originalType == OriginalType.TIMESTAMP_MICROS) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              column.putLong(i, dictionary.decodeToLong(dictionaryIds.getDictId(i)));
+            }
+          }
+        } else if (originalType == OriginalType.TIMESTAMP_MILLIS) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              column.putLong(i,
+                DateTimeUtils.fromMillis(dictionary.decodeToLong(dictionaryIds.getDictId(i))));
+            }
+          }
+        } else {
+          throw constructConvertNotSupportedException(descriptor, column);
+        }
+        break;
+
+      case FLOAT:
+        for (int i = rowId; i < rowId + num; ++i) {
+          if (!column.isNullAt(i)) {
+            column.putFloat(i, dictionary.decodeToFloat(dictionaryIds.getDictId(i)));
+          }
+        }
+        break;
+
+      case DOUBLE:
+        for (int i = rowId; i < rowId + num; ++i) {
+          if (!column.isNullAt(i)) {
+            column.putDouble(i, dictionary.decodeToDouble(dictionaryIds.getDictId(i)));
+          }
+        }
+        break;
+      case INT96:
+        if (column.dataType() == DataTypes.TimestampType) {
+          if (!shouldConvertTimestamps()) {
+            for (int i = rowId; i < rowId + num; ++i) {
+              if (!column.isNullAt(i)) {
+                Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+                column.putLong(i, ParquetRowConverter.binaryToSQLTimestamp(v));
+              }
+            }
+          } else {
+            for (int i = rowId; i < rowId + num; ++i) {
+              if (!column.isNullAt(i)) {
+                Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+                long rawTime = ParquetRowConverter.binaryToSQLTimestamp(v);
+                long adjTime = DateTimeUtils.convertTz(rawTime, convertTz, UTC);
+                column.putLong(i, adjTime);
+              }
+            }
+          }
+        } else {
+          throw constructConvertNotSupportedException(descriptor, column);
+        }
+        break;
+      case BINARY:
+        // TODO: this is incredibly inefficient as it blows up the dictionary right here. We
+        // need to do this better. We should probably add the dictionary data to the ColumnVector
+        // and reuse it across batches. This should mean adding a ByteArray would just update
+        // the length and offset.
+        for (int i = rowId; i < rowId + num; ++i) {
+          if (!column.isNullAt(i)) {
+            Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+            column.putByteArray(i, v.getBytes());
+          }
+        }
+        break;
+      case FIXED_LEN_BYTE_ARRAY:
+        // DecimalType written in the legacy mode
+        if (DecimalType.is32BitDecimalType(column.dataType())) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+              column.putInt(i, (int) ParquetRowConverter.binaryToUnscaledLong(v));
+            }
+          }
+        } else if (DecimalType.is64BitDecimalType(column.dataType())) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+              column.putLong(i, ParquetRowConverter.binaryToUnscaledLong(v));
+            }
+          }
+        } else if (DecimalType.isByteArrayDecimalType(column.dataType())) {
+          for (int i = rowId; i < rowId + num; ++i) {
+            if (!column.isNullAt(i)) {
+              Binary v = dictionary.decodeToBinary(dictionaryIds.getDictId(i));
+              column.putByteArray(i, v.getBytes());
+            }
+          }
+        } else {
+          throw constructConvertNotSupportedException(descriptor, column);
+        }
+        break;
+
+      default:
+        throw new UnsupportedOperationException(
+                "Unsupported type: " + descriptor.getPrimitiveType().getPrimitiveTypeName());
+    }
+  }
+
+  /**
+   * For all the read*Batch functions, reads `num` values from this columnReader into column. It
+   * is guaranteed that num is smaller than the number of values left in the current page.
+   */
+
+  private void readBooleanBatch(int rowId, int num, WritableColumnVector column)
+          throws IOException {
+    if (column.dataType() != DataTypes.BooleanType) {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+    defColumnRef.readBooleans(
+            num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+  }
+
+  private void readIntBatch(int rowId, int num, WritableColumnVector column) throws IOException {
+    // This is where we implement support for the valid type conversions.
+    // TODO: implement remaining type conversions
+    if (column.dataType() == DataTypes.IntegerType || column.dataType() == DataTypes.DateType ||
+            DecimalType.is32BitDecimalType(column.dataType())) {
+      defColumnRef.readIntegers(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else if (column.dataType() == DataTypes.ByteType) {
+      defColumnRef.readBytes(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else if (column.dataType() == DataTypes.ShortType) {
+      defColumnRef.readShorts(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  private void readLongBatch(int rowId, int num, WritableColumnVector column) throws IOException {
+    // This is where we implement support for the valid type conversions.
+    if (column.dataType() == DataTypes.LongType ||
+            DecimalType.is64BitDecimalType(column.dataType()) ||
+            originalType == OriginalType.TIMESTAMP_MICROS) {
+      defColumnRef.readLongs(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else if (originalType == OriginalType.TIMESTAMP_MILLIS) {
+      for (int i = 0; i < num; i++) {
+        if (defColumnRef.readInteger() == maxDefLevel) {
+          column.putLong(rowId + i, DateTimeUtils.fromMillis(dataColumn.readLong()));
+        } else {
+          column.putNull(rowId + i);
+        }
+      }
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  private void readFloatBatch(int rowId, int num, WritableColumnVector column) throws IOException {
+    // This is where we implement support for the valid type conversions.
+    // TODO: support implicit cast to double?
+    if (column.dataType() == DataTypes.FloatType) {
+      defColumnRef.readFloats(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  private void readDoubleBatch(int rowId, int num, WritableColumnVector column) throws IOException {
+    // This is where we implement support for the valid type conversions.
+    // TODO: implement remaining type conversions
+    if (column.dataType() == DataTypes.DoubleType) {
+      defColumnRef.readDoubles(
+              num, column, rowId, maxDefLevel, (VectorizedValuesReader) dataColumn);
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  private void readBinaryBatch(int rowId, int num, WritableColumnVector column) throws IOException {
+    // This is where we implement support for the valid type conversions.
+    // TODO: implement remaining type conversions
+    VectorizedValuesReader data = (VectorizedValuesReader) dataColumn;
+    if (column.dataType() == DataTypes.StringType || column.dataType() == DataTypes.BinaryType
+            || DecimalType.isByteArrayDecimalType(column.dataType())) {
+      defColumnRef.readBinarys(num, column, rowId, maxDefLevel, data);
+    } else if (column.dataType() == DataTypes.TimestampType) {
+      if (!shouldConvertTimestamps()) {
+        for (int i = 0; i < num; i++) {
+          if (dataColumnRef.readInteger() == maxDefLevel) {
+            // Read 12 bytes for INT96
+            long rawTime = ParquetRowConverter.binaryToSQLTimestamp(data.readBinary(12));
+            column.putLong(rowId + i, rawTime);
+          } else {
+            column.putNull(rowId + i);
+          }
+        }
+      } else {
+        for (int i = 0; i < num; i++) {
+          if (defColumnRef.readInteger() == maxDefLevel) {
+            // Read 12 bytes for INT96
+            long rawTime = ParquetRowConverter.binaryToSQLTimestamp(data.readBinary(12));
+            long adjTime = DateTimeUtils.convertTz(rawTime, convertTz, UTC);
+            column.putLong(rowId + i, adjTime);
+          } else {
+            column.putNull(rowId + i);
+          }
+        }
+      }
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  private void readFixedLenByteArrayBatch(
+          int rowId,
+          int num,
+          WritableColumnVector column,
+          int arrayLen) {
+    VectorizedValuesReader data = (VectorizedValuesReader) dataColumn;
+    // This is where we implement support for the valid type conversions.
+    // TODO: implement remaining type conversions
+    if (DecimalType.is32BitDecimalType(column.dataType())) {
+      for (int i = 0; i < num; i++) {
+        if (defColumnRef.readInteger() == maxDefLevel) {
+          column.putInt(rowId + i,
+                  (int) ParquetRowConverter.binaryToUnscaledLong(data.readBinary(arrayLen)));
+        } else {
+          column.putNull(rowId + i);
+        }
+      }
+    } else if (DecimalType.is64BitDecimalType(column.dataType())) {
+      for (int i = 0; i < num; i++) {
+        if (defColumnRef.readInteger() == maxDefLevel) {
+          column.putLong(rowId + i,
+                  ParquetRowConverter.binaryToUnscaledLong(data.readBinary(arrayLen)));
+        } else {
+          column.putNull(rowId + i);
+        }
+      }
+    } else if (DecimalType.isByteArrayDecimalType(column.dataType())) {
+      for (int i = 0; i < num; i++) {
+        if (defColumnRef.readInteger() == maxDefLevel) {
+          column.putByteArray(rowId + i, data.readBinary(arrayLen).getBytes());
+        } else {
+          column.putNull(rowId + i);
+        }
+      }
+    } else {
+      throw constructConvertNotSupportedException(descriptor, column);
+    }
+  }
+
+  protected void readPage() {
+    DataPage page = pageReader.readPage();
+    // TODO: Why is this a visitor?
+    page.accept(new DataPage.Visitor<Void>() {
+      @Override
+      public Void visit(DataPageV1 dataPageV1) {
+        try {
+          readPageV1(dataPageV1);
+          return null;
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
+      }
+
+      @Override
+      public Void visit(DataPageV2 dataPageV2) {
+        try {
+          readPageV2(dataPageV2);
+          return null;
+        } catch (IOException e) {
+          throw new RuntimeException(e);
+        }
+      }
+    });
   }
 
   /**
@@ -90,7 +593,7 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
       // isCurrentPageDictionaryEncoded re-assignment by readPage method.
       if (isCurrentPageDictionaryEncoded) {
         // If isCurrentPageDictionaryEncoded is true, dataType must be INT32, call skipIntegers.
-        ((SkippableVectorizedRleValuesReader)defColumn)
+        ((SkippableVectorizedRleValuesReader)defColumnRef)
           .skipIntegers(num, maxDefLevel, (SkippableVectorizedValuesReader) dataColumn);
       } else {
         // isCurrentPageDictionaryEncoded is false, call skip by descriptor.getType(), this type
@@ -262,7 +765,6 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
    * This method refer to initDataReader in VectorizedColumnReader,
    * just modified the assignment of dataColumn.
    */
-  @Override
   protected void initDataReader(Encoding dataEncoding, ByteBufferInputStream in)
     throws IOException {
     this.endOfPageValueCount = valuesRead + pageValueCount;
@@ -304,7 +806,6 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
    * modified the assignment of defColumn and remove assignment to
    * repetitionLevelColumn & definitionLevelColumn because they are useless.
    */
-  @Override
   protected void readPageV1(DataPageV1 page) throws IOException {
     this.pageValueCount = page.getValueCount();
     ValuesReader rlReader = page.getRlEncoding().getValuesReader(descriptor, REPETITION_LEVEL);
@@ -315,10 +816,10 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
       throw new UnsupportedOperationException("Unsupported encoding: " + page.getDlEncoding());
     }
     int bitWidth = BytesUtils.getWidthFromMaxInt(descriptor.getMaxDefinitionLevel());
-    this.defColumn = new SkippableVectorizedRleValuesReader(bitWidth);
+    this.defColumnRef = new SkippableVectorizedRleValuesReader(bitWidth);
     // defColumnRef reference defColumn and type is SkippableVectorizedRleValuesReader
-    this.defColumnRef = (SkippableVectorizedRleValuesReader)this.defColumn;
-    dlReader = this.defColumn;
+    this.defColumnRef = (SkippableVectorizedRleValuesReader)this.defColumnRef;
+    dlReader = this.defColumnRef;
     try {
       BytesInput bytes = page.getBytes();
       ByteBufferInputStream in = bytes.toInputStream();
@@ -335,17 +836,16 @@ public class SkippableVectorizedColumnReader extends VectorizedColumnReader {
    * modified the assignment of defColumn and remove assignment to
    * repetitionLevelColumn & definitionLevelColumn because they are useless.
    */
-  @Override
   protected void readPageV2(DataPageV2 page) throws IOException {
     this.pageValueCount = page.getValueCount();
 
     int bitWidth = BytesUtils.getWidthFromMaxInt(descriptor.getMaxDefinitionLevel());
     // do not read the length from the stream. v2 pages handle dividing the page bytes.
-    this.defColumn = new SkippableVectorizedRleValuesReader(bitWidth, false);
-    this.defColumn.initFromPage(
+    this.defColumnRef = new SkippableVectorizedRleValuesReader(bitWidth, false);
+    this.defColumnRef.initFromPage(
             this.pageValueCount, page.getDefinitionLevels().toInputStream());
     // defColumnRef reference defColumn and type is SkippableVectorizedRleValuesReader
-    this.defColumnRef = (SkippableVectorizedRleValuesReader) this.defColumn;
+    this.defColumnRef = (SkippableVectorizedRleValuesReader) this.defColumnRef;
     try {
       initDataReader(page.getDataEncoding(), page.getData().toInputStream());
     } catch (IOException e) {
